+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-- Support for Azure Service Bus Queue.
+- Support for Azure Service Bus Queue. https://github.com/Azure/azure-mcp/pull/144
 
 ### Breaking Changes